Remove custom font
In the current version of The Lounge, this font doesn't load without extra hoops, which ends up being Times New Roman (at least on Mac, haven't checked somewhere else).

Also, now that the main repo uses native font stack, it is sensible to rely on it. This custom font could be re-added once support for custom fonts in themes is added to The Lounge if you want.